\name{print.NSM3Ch5p}
\alias{print.NSM3Ch5p}
\alias{print.NSM3Ch5c}
\alias{print.NSM3Ch6p}
\alias{print.NSM3Ch6c}
\alias{print.NSM3Ch7p}
\alias{print.NSM3Ch7c}
\alias{print.NSM3Ch6MCp}
\alias{print.NSM3Ch6MCc}
\alias{print.NSM3Ch7MCp}
\alias{print.NSM3Ch7MCc}

\title{
Methods to control displayed output of NSM3 tests.
}
\description{
These methods are used to display the list output from the functions used to perform the various nonparametric statistical procedures in the NSM3 package.
}
\usage{
\method{print}{NSM3Ch5p}(x, ...)
}
\arguments{
  \item{x}{The list object returned by a procedure in the NSM3 package.}
  \item{...}{Other options to be specified.}
}
\value{
The exact wording of the displayed output will vary depending on the setting. For example two sample procedures and k-sample procedures will be worded in a slightly different manner.
}
\author{
Grant Schneider
}
\keyword{print}
\keyword{NSM3}


